Split Chaquopy into independent XML, Compose, and Litho template sources
