+++ utilities/templates-impl/src/main/java/com/itsaky/androidide/templates/impl/TemplateProviderImpl.kt
@@ -37,9 +37,9 @@ import com.itsaky.androidide.templates.impl.androidstudio.activities.viewModelAc
 import com.itsaky.androidide.templates.impl.androidstudio.activities.xrActivity.xrActivityTemplate
 import com.itsaky.androidide.templates.impl.basicCpp.basicCppProject
 import com.itsaky.androidide.templates.impl.bottomNavActivity.bottomNavActivityProject
-import com.itsaky.androidide.templates.impl.chaquopy.chaquopyComposeDemoProject
-import com.itsaky.androidide.templates.impl.chaquopy.chaquopyLithoDemoProject
-import com.itsaky.androidide.templates.impl.chaquopy.chaquopyXmlDemoProject
+import com.itsaky.androidide.templates.impl.chaquopy.compose.chaquopyComposeDemoProject
+import com.itsaky.androidide.templates.impl.chaquopy.litho.chaquopyLithoDemoProject
+import com.itsaky.androidide.templates.impl.chaquopy.xml.chaquopyXmlDemoProject
 import com.itsaky.androidide.templates.impl.composeActivity.composeActivityProject
 import com.itsaky.androidide.templates.impl.emptyActivity.emptyActivityProject
 import com.itsaky.androidide.templates.impl.lithoClassic.lithoClassicProject
